Tests: print contents of *EGLOT ...* buffers in batch mode.
Useful for the CI on github. To be able to see more of the context of
a failure.

* eglot.el (eglot--server-created-hook): New. Called when an instance
of `eglot-lsp-server' is created as part of the "connect to server"
flow. This hook is internal only right now: it has been added for the
test suite. The existing `eglot-connect-hook' and
`eglot-server-initialized-hook' are run once a connection was
successfully established. They do not capture server instances that
failed to be started. And the test suite better also capture those,
too, such that their process output can be dumped.
(eglot--connect): Use it.

* eglot-tests.el (eglot--call-with-fixture): 1) use the new
`eglot--server-created-hook' rather than `eglot-connect-hook';
2) Dump the contents of the *EGLOT ...* buffers when run in
`noninteractive' (i.e. batch) mode.
(eglot--cleanup-after-test): New auxiliary function. Extracted
verbatim out of `eglot--call-with-fixture` in order to lower the
latter's LOC.